+++ cocos/platform/win32/FileUtils-win32.cpp
@@ -48,8 +48,8 @@ static std::string s_resourcePath = "";
 // D:\aaa\bbb\ccc\ddd\abc.txt --> D:/aaa/bbb/ccc/ddd/abc.txt
 static inline std::string convertPathFormatToUnixStyle(const std::string &path) {
     std::string ret = path;
-    size_t      len = ret.length();
-    for (size_t i = 0; i < len; ++i) {
+    int         len = ret.length();
+    for (int i = 0; i < len; ++i) {
         if (ret[i] == '\\') {
             ret[i] = '/';
         }
@@ -67,8 +67,7 @@ static void _checkPath() {
         WCHAR *pUtf16DirEnd = wcsrchr(pUtf16ExePath, L'\\');
 
         char utf8ExeDir[CC_MAX_PATH] = {0};
-        int  nNum                    = WideCharToMultiByte(CP_UTF8, 0, pUtf16ExePath, static_cast<int>(pUtf16DirEnd - pUtf16ExePath + 1),
-                                       utf8ExeDir, static_cast<int>(sizeof(utf8ExeDir)), nullptr, nullptr);
+        int  nNum                    = WideCharToMultiByte(CP_UTF8, 0, pUtf16ExePath, pUtf16DirEnd - pUtf16ExePath + 1, utf8ExeDir, sizeof(utf8ExeDir), nullptr, nullptr);
 
         s_resourcePath = convertPathFormatToUnixStyle(utf8ExeDir);
     }
